portfolio/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-grappelli==2.15.3
django-modeltranslation==0.17.3
python-dotenv==0.19.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
